+++ uae4all_gp2x_0.7.2a/src/include/cia.h
@@ -14,9 +14,9 @@ extern void CIA_handler (void);
 extern void diskindex_handler (void);
 
 extern void rethink_cias (void);
-extern unsigned int ciaaicr,ciaaimask,ciabicr,ciabimask;
-extern unsigned int ciaacra,ciaacrb,ciabcra,ciabcrb;
-extern unsigned int ciabpra;
-extern unsigned long ciaata,ciaatb,ciabta,ciabtb;
-extern unsigned long ciaatod,ciabtod,ciaatol,ciabtol,ciaaalarm,ciabalarm;
+extern uint32_t ciaaicr,ciaaimask,ciabicr,ciabimask;
+extern uint32_t ciaacra,ciaacrb,ciabcra,ciabcrb;
+extern uint32_t ciabpra;
+extern uint32_t ciaata,ciaatb,ciabta,ciabtb;
+extern uint32_t ciaatod,ciabtod,ciaatol,ciabtol,ciaaalarm,ciabalarm;
 extern int ciaatlatch,ciabtlatch;
